Command Line entry point and parser.
Implementation of a command line parser using the argparse4j.
It has two mode of operation, build to create the inverted index and
search to start the search module.

The AppSearch and AppIndex class are used as the entry point for each
mode by the parser.

Currently those two classes are stubs and only print a success message.